s01e03/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
langchain==1.2.11
langchain-openai==1.1.11
langgraph==1.1.0
python-dotenv==1.2.2
